refactor: Rename `lib/plugins/aws/lib/updateStack.js`
