+++ lib/plugins/aws/deploy/index.js
@@ -14,7 +14,7 @@ const createStack = require('./lib/createStack');
 const cleanupS3Bucket = require('./lib/cleanupS3Bucket');
 const uploadArtifacts = require('./lib/uploadArtifacts');
 const validateTemplate = require('./lib/validateTemplate');
-const updateStack = require('../lib/updateStack');
+const updateStack = require('../lib/update-stack');
 const ensureValidBucketExists = require('./lib/ensure-valid-bucket-exists');
 const path = require('path');
 const { style, log, progress, writeText } = require('@serverless/utils/log');
