+++ lib/plugins/aws/lib/update-stack.js
