+++ lib/plugins/aws/rollback.js
@@ -1,7 +1,7 @@
 'use strict';
 const validate = require('./lib/validate');
 const setBucketName = require('./lib/set-bucket-name');
-const updateStack = require('./lib/updateStack');
+const updateStack = require('./lib/update-stack');
 const monitorStack = require('./lib/monitor-stack');
 const waitForChangeSetCreation = require('./lib/wait-for-change-set-creation');
 const getCreateChangeSetParams = require('./lib/get-create-change-set-params');
